pdf_generation_svc/requirements.txt
pika
reportlab
grpcio
grpcio-tools